+++ permission/src/androidMain/kotlin/tech/kotlinlang/permission/AndroidPermissionHelper.kt
@@ -1,16 +1,23 @@
 package tech.kotlinlang.permission
 
 import android.Manifest
+import android.content.Context
 import android.content.pm.PackageManager
 import android.os.Build
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import tech.kotlinlang.permission.result.LocationPermissionResult
 import tech.kotlinlang.permission.result.NotificationPermissionResult
+import androidx.core.content.edit
 
 class AndroidPermissionHelper(
     private val permissionInitiation: PermissionInitiation
 ) : PermissionHelper {
+
+    private val sharedPref by lazy {
+        val activity = permissionInitiation.getActivity()
+        activity.getSharedPreferences("PermissionPrefs", Context.MODE_PRIVATE)
+    }
 
     override suspend fun <T> checkIsPermissionGranted(permission: Permission<T>): T {
         @Suppress("UNCHECKED_CAST")
@@ -33,12 +40,13 @@ class AndroidPermissionHelper(
             return NotificationPermissionResult.Granted
         }
         val activity = permissionInitiation.getActivity()
+        val isUsed = sharedPref.getBoolean("notification.permission", false)
         return if (ContextCompat.checkSelfPermission(
                 activity, Manifest.permission.POST_NOTIFICATIONS
             ) == PackageManager.PERMISSION_GRANTED
         ) {
             NotificationPermissionResult.Granted
-        } else if (!ActivityCompat.shouldShowRequestPermissionRationale(
+        } else if (isUsed && !ActivityCompat.shouldShowRequestPermissionRationale(
                 activity,
                 Manifest.permission.POST_NOTIFICATIONS,
             )
@@ -54,11 +62,13 @@ class AndroidPermissionHelper(
             return NotificationPermissionResult.Granted
         }
         permissionInitiation.requestPermission(Manifest.permission.POST_NOTIFICATIONS)
+        sharedPref.edit { putBoolean("notification.permission", true) }
         return checkNotificationPermission()
     }
 
     private fun checkLocationPermission(): LocationPermissionResult {
         val activity = permissionInitiation.getActivity()
+        val isUsed = sharedPref.getBoolean("location.permission", false)
         return if (ContextCompat.checkSelfPermission(
                 activity,
                 Manifest.permission.ACCESS_FINE_LOCATION
@@ -71,7 +81,7 @@ class AndroidPermissionHelper(
             ) == PackageManager.PERMISSION_GRANTED
         ) {
             LocationPermissionResult.Granted.Approximate
-        } else if (!ActivityCompat.shouldShowRequestPermissionRationale(
+        } else if (isUsed && !ActivityCompat.shouldShowRequestPermissionRationale(
                 activity,
                 Manifest.permission.ACCESS_COARSE_LOCATION
             )
@@ -84,6 +94,7 @@ class AndroidPermissionHelper(
 
     private suspend fun requestLocationPermission(): LocationPermissionResult {
         permissionInitiation.requestPermission(Manifest.permission.ACCESS_FINE_LOCATION)
+        sharedPref.edit { putBoolean("location.permission", true) }
         return checkLocationPermission()
     }
 }