+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Fix
+- Fix Not Allowed issue while checking permission for first time in Android
+
 ## [0.1.0] - 2025-04-17
 
 ### Added